Trait thinlinelib::value_parser::ValueParser

pub trait ValueParser {
    fn get_bool(&self, keys: &[&str], default: bool) -> bool;

    fn get_str(&self, keys: &[&str]) -> Option<&str>;

    fn get_str_or_file_content(
        &self, 
        keys: &[&str], 
        base_path: &PathBuf, 
        file_ext: &str
    ) -> Option<&str>;

    fn get_str_vec(&self, keys: &[&str]) -> Option<Vec<&str>>;
}

Required Methods

fn get_bool(&self, keys: &[&str], default: bool) -> bool

Reads a bool at the given config path.

fn get_str(&self, keys: &[&str]) -> Option<&str>

Reads a string at the given config path.

fn get_str_or_file_content(
&self,
keys: &[&str],
base_path: &PathBuf,
file_ext: &str
) -> Option<&str>

Reads a string or the content of the file at the given path.

fn get_str_vec(&self, keys: &[&str]) -> Option<Vec<&str>>

Reads a string vector at the given config path.